SOW-24: rename hook and fix margin
