+++ src/main.tsx
@@ -6,10 +6,12 @@ import { FormPageHoC } from "./pages/FormPageHoc";
 import { validateDescription } from "./pages/description-page/description-validation";
 import { validateTimetableStage } from "./pages/stage-page/validate-stage-page";
 import { validateTitle } from "./pages/title-page/validate-title-page";
-import { stages } from "./pages/useSequence";
+import { stages } from "./pages/use-sequence";
 import { Page } from "./routes/Page";
 import { Root } from "./routes/Root";
 import { PageRoute } from "./routes/routes";
+
+import "./main.css";
 
 const router = createBrowserRouter(
   [
